Optimized the sketch and query component

+++ dsserver/src/main/java/rptu/thesis/npham/dsserver/model/similarity/Measure.java
@@ -1,6 +1,6 @@
 package rptu.thesis.npham.dsserver.model.similarity;
 
-public record Measure(MeasureType measures, double score, double weight) {
+public record Measure(MeasureType measures, double score, int weight) {
     @Override
     public String toString() {
         return measures + "= " + score;
